+++ onnxruntime/core/providers/nnapi/nnapi_builtin/builders/helper.h
@@ -3,7 +3,7 @@
 //
 #pragma once
 
-#include <core/common/common.h>
+#include <core/graph/graph.h>
 #include <string>
 
 #include "core/providers/nnapi/nnapi_builtin/nnapi_lib/NeuralNetworksTypes.h"
@@ -61,3 +61,22 @@ inline std::string GetErrorCause(int error_code) {
       return "Unknown error code: " + std::to_string(error_code);
   }
 }
+
+/**
+ * Wrapping onnxruntime::Node for retrieving attribute values
+ */
+class NodeAttrHelper {
+ public:
+  NodeAttrHelper(const onnxruntime::Node& proto);
+
+  float Get(const std::string& key, float def_val) const;
+  int32_t Get(const std::string& key, int32_t def_val) const;
+  std::vector<float> Get(const std::string& key, const std::vector<float>& def_val) const;
+  std::vector<int32_t> Get(const std::string& key, const std::vector<int32_t>& def_val) const;
+  std::string Get(const std::string& key, const std::string& def_val) const;
+
+  bool HasAttr(const std::string& key) const;
+
+ private:
+  const onnxruntime::NodeAttributes& node_attributes_;
+};
