--- onnxruntime/core/providers/nnapi/nnapi_builtin/builders/node_attr_helper.h
@@ -1,44 +1,0 @@
-//
-// Created by daquexian on 8/3/18.
-//
-
-#pragma once
-
-#include <core/graph/graph.h>
-#include <onnx/onnx_pb.h>
-#include <string>
-
-/**
- * Wrapping onnx::NodeProto for retrieving attribute values
- */
-class NodeAttrHelper {
- public:
-  NodeAttrHelper(const ONNX_NAMESPACE::NodeProto& proto);
-
-  float Get(const std::string& key, float def_val);
-  int32_t Get(const std::string& key, int32_t def_val);
-  std::vector<float> Get(const std::string& key, const std::vector<float>& def_val);
-  std::vector<int32_t> Get(const std::string& key, const std::vector<int32_t>& def_val);
-  std::string Get(const std::string& key, const std::string& def_val);
-
-  bool HasAttr(const std::string& key);
-
- private:
-  const ONNX_NAMESPACE::NodeProto& node_;
-};
-
-class GraphNodeAttrHelper {
- public:
-  GraphNodeAttrHelper(const onnxruntime::Node& proto);
-
-  float Get(const std::string& key, float def_val) const;
-  int32_t Get(const std::string& key, int32_t def_val) const;
-  std::vector<float> Get(const std::string& key, const std::vector<float>& def_val) const;
-  std::vector<int32_t> Get(const std::string& key, const std::vector<int32_t>& def_val) const;
-  std::string Get(const std::string& key, const std::string& def_val) const;
-
-  bool HasAttr(const std::string& key) const;
-
- private:
-  const onnxruntime::NodeAttributes& node_attributes_;
-};
