+++ onnxruntime/core/providers/nnapi/nnapi_builtin/builders/op_builder.h
@@ -14,24 +14,12 @@ class IOpBuilder {
 
   // Check if an operator is supported
   virtual bool IsOpSupported(ModelBuilder& model_builder,
-                             const ONNX_NAMESPACE::NodeProto& node) = 0;
-
-  virtual bool IsOpSupported(ModelBuilder& model_builder,
                              const onnxruntime::Node& node) = 0;
 
   // Check if the initializers of this operator need preprocess
   // which will not be copied
   virtual void AddInitializersToSkip(ModelBuilder& model_builder,
-                                     const ONNX_NAMESPACE::NodeProto& node) = 0;
-
-  // Check if the initializers of this operator need preprocess
-  // which will not be copied
-  virtual void AddInitializersToSkip(ModelBuilder& model_builder,
                                      const onnxruntime::Node& node) = 0;
-
-  // Add the operator to NNAPI model
-  virtual void AddToModelBuilder(ModelBuilder& model_builder,
-                                 const ONNX_NAMESPACE::NodeProto& node) = 0;
 
   // Add the operator to NNAPI model
   virtual void AddToModelBuilder(ModelBuilder& model_builder,
