+++ onnxruntime/core/providers/nnapi/nnapi_builtin/nnapi_execution_provider.cc
@@ -53,41 +53,14 @@ NnapiExecutionProvider::GetCapability(const onnxruntime::GraphViewer& graph_view
     }
   }
 
-  // TODO, switch to use graph instead of model
-  // This method is based on that of TRT EP
-  // Construct modelproto from graph
-  onnxruntime::Model model(graph_view.Name(), true, ModelMetaData(),
-                           PathString(),
-                           IOnnxRuntimeOpSchemaRegistryList(),
-                           graph_view.DomainToVersionMap(),
-                           std::vector<ONNX_NAMESPACE::FunctionProto>(),
-                           *GetLogger());
   std::unordered_set<std::string> all_node_inputs;
-  onnxruntime::Graph& graph_build = model.MainGraph();
   for (const auto& node : graph_view.Nodes()) {
-    std::vector<onnxruntime::NodeArg*> inputs, outputs;
     for (auto* input : node.InputDefs()) {
-      auto& n_input = graph_build.GetOrCreateNodeArg(input->Name(), input->TypeAsProto());
-      inputs.push_back(&n_input);
       all_node_inputs.insert(input->Name());
     }
-    for (auto* output : node.OutputDefs()) {
-      auto& n_output = graph_build.GetOrCreateNodeArg(output->Name(), output->TypeAsProto());
-      outputs.push_back(&n_output);
-    }
-    graph_build.AddNode(node.Name(), node.OpType(), node.Description(), inputs, outputs, &node.GetAttributes(), node.Domain());
-  }
-  //Add initializer to graph
-  const auto& init_tensors = graph_view.GetAllInitializedTensors();
-  for (const auto& tensor : init_tensors) {
-    graph_build.AddInitializedTensor(*(tensor.second));
   }
 
-  ORT_ENFORCE(graph_build.Resolve().IsOK());
-  ONNX_NAMESPACE::ModelProto model_proto = model.ToProto();
-  model_proto.set_ir_version(ONNX_NAMESPACE::Version::IR_VERSION);
-
-  nnapi::ModelBuilder builder(model_proto, graph_view);
+  nnapi::ModelBuilder builder(graph_view);
   const auto supported_nodes_vector = builder.GetSupportedNodes();
 
   // Find inputs, initializers and outputs for each supported subgraph
@@ -222,13 +195,10 @@ common::Status NnapiExecutionProvider::Compile(const std::vector<onnxruntime::No
     onnxruntime::Model model(graph_body.Name(), true, ModelMetaData(), PathString(),
                              IOnnxRuntimeOpSchemaRegistryList(), graph_body.DomainToVersionMap(),
                              std::vector<ONNX_NAMESPACE::FunctionProto>(), *GetLogger());
-    ONNX_NAMESPACE::ModelProto model_proto = model.ToProto();
-    *(model_proto.mutable_graph()) = graph_body.ToGraphProto();
-    model_proto.set_ir_version(ONNX_NAMESPACE::Version::IR_VERSION);
 
     {
       onnxruntime::GraphViewer graph_viewer(graph_body);
-      nnapi::ModelBuilder builder(model_proto, graph_viewer);
+      nnapi::ModelBuilder builder(graph_viewer);
       builder.SetUseNCHW(false);
       builder.SetUseFp16(false);
       std::unique_ptr<nnapi::Model> nnapi_model = builder.Compile();
