+++ onnxruntime/core/providers/nnapi/nnapi_builtin/nnapi_lib/NeuralNetworksWrapper.h
@@ -17,8 +17,6 @@
 #ifndef ANDROID_ML_NN_RUNTIME_NEURAL_NETWORKS_WRAPPER_H
 #define ANDROID_ML_NN_RUNTIME_NEURAL_NETWORKS_WRAPPER_H
 #include "nnapi_implementation.h"
-#include <math.h>
-#include <string>
 #include <vector>
 #include <numeric>
